+++ internal/stremio/transformer/stream_extractor.go
@@ -133,12 +133,12 @@ type StreamExtractorResult struct {
 	File      StreamExtractorResultFile
 	Hash      string
 	IsPrivate bool
-	Raw       StreamExtractorResultRaw `expr:"-"`
+	Raw       StreamExtractorResultRaw
 	Season    int
 	Seeders   int
 	Store     StreamExtractorResultStore
 	TTitle    string `expr:"-"`
-	Indexer   string
+	Indexer   string `expr:"-"`
 }
 
 var language_to_code = map[string]string{
@@ -149,6 +149,7 @@ var language_to_code = map[string]string{
 
 	"english":    "en",
 	"🇬🇧":         "en",
+	"🇺🇸":         "en",
 	"japanese":   "ja",
 	"🇯🇵":         "ja",
 	"russian":    "ru",
@@ -157,6 +158,7 @@ var language_to_code = map[string]string{
 	"🇮🇹":         "it",
 	"portuguese": "pt",
 	"🇵🇹":         "pt",
+	"🇧🇷":         "pt",
 	"spanish":    "es",
 	"🇪🇸":         "es",
 	"latino":     "es-419",