+++ internal/worker/torznab_indexer_syncer.go
@@ -7,6 +7,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/MunifTanjim/stremthru/internal/anidb"
 	"github.com/MunifTanjim/stremthru/internal/imdb_title"
 	"github.com/MunifTanjim/stremthru/internal/torrent_info"
 	"github.com/MunifTanjim/stremthru/internal/torrent_stream"
@@ -14,6 +15,7 @@ import (
 	torznab_indexer "github.com/MunifTanjim/stremthru/internal/torznab/indexer"
 	torznab_indexer_syncinfo "github.com/MunifTanjim/stremthru/internal/torznab/indexer/syncinfo"
 	"github.com/MunifTanjim/stremthru/internal/util"
+	"github.com/alitto/pond/v2"
 )
 
 func InitTorznabIndexerSyncerWorker(conf *WorkerConfig) *Worker {
@@ -88,36 +90,71 @@ func InitTorznabIndexerSyncerWorker(conf *WorkerConfig) *Worker {
 				for i := range items {
 					item := &items[i]
 
+					queryMeta := indexerQueryMeta{
+						titles: []string{},
+					}
+
 					nsid, err := torrent_stream.NormalizeStreamId(item.SId)
 					if err != nil {
 						log.Error("failed to normalize stream ID", "error", err, "sid", item.SId)
 						continue
 					}
-
-					it, err := imdb_title.Get(nsid.Id)
-					if err != nil {
-						log.Error("failed to get IMDB title", "error", err, "imdb_id", nsid.Id)
-						continue
-					}
-					if it == nil {
-						log.Warn("IMDB title not found", "imdb_id", nsid.Id)
-						continue
-					}
-
-					queryMeta := indexerQueryMeta{
-						titles: []string{},
-					}
-
-					queryMeta.titles = append(queryMeta.titles, it.Title)
-					if it.OrigTitle != "" && it.OrigTitle != it.Title {
-						queryMeta.titles = append(queryMeta.titles, it.OrigTitle)
-					}
-					if it.Year > 0 {
-						queryMeta.year = it.Year
-					}
-					if nsid.IsSeries() {
-						queryMeta.season = util.SafeParseInt(nsid.Season, 0)
-						queryMeta.ep = util.SafeParseInt(nsid.Episode, 0)
+					if nsid.IsAnime {
+						if aniEp := util.SafeParseInt(nsid.Episode, -1); aniEp != -1 {
+							tvdbMaps, err := anidb.GetTVDBEpisodeMaps(nsid.Id, false)
+							if err != nil {
+								log.Error("failed to get AniDB-TVDB episode maps", "error", err, "anidb_id", nsid.Id)
+								continue
+							}
+							if epMap := tvdbMaps.GetByAnidbEpisode(aniEp); epMap != nil {
+								ep := epMap.GetTMDBEpisode(aniEp)
+								titles, err := anidb.GetTitlesByIds([]string{nsid.Id})
+								if err != nil {
+									log.Error("failed to get AniDB titles", "error", err, "anidb_id", nsid.Id)
+									continue
+								}
+								if len(titles) == 0 {
+									log.Warn("AniDB title not found", "anidb_id", nsid.Id)
+									continue
+								}
+								queryMeta.titles = make([]string, 0, len(titles))
+								queryMeta.season = epMap.TVDBSeason
+								queryMeta.ep = ep
+								seenTitle := util.NewSet[string]()
+								for i := range titles {
+									title := &titles[i]
+									if seenTitle.Has(title.Value) {
+										continue
+									}
+									seenTitle.Add(title.Value)
+									queryMeta.titles = append(queryMeta.titles, title.Value)
+									if queryMeta.year == 0 && title.Year != "" {
+										queryMeta.year = util.SafeParseInt(title.Year, 0)
+									}
+								}
+							}
+						}
+					} else {
+						it, err := imdb_title.Get(nsid.Id)
+						if err != nil {
+							log.Error("failed to get IMDB title", "error", err, "imdb_id", nsid.Id)
+							continue
+						}
+						if it == nil {
+							log.Warn("IMDB title not found", "imdb_id", nsid.Id)
+							continue
+						}
+						queryMeta.titles = append(queryMeta.titles, it.Title)
+						if it.OrigTitle != "" && it.OrigTitle != it.Title {
+							queryMeta.titles = append(queryMeta.titles, it.OrigTitle)
+						}
+						if it.Year > 0 {
+							queryMeta.year = it.Year
+						}
+						if nsid.IsSeries() {
+							queryMeta.season = util.SafeParseInt(nsid.Season, 0)
+							queryMeta.ep = util.SafeParseInt(nsid.Episode, 0)
+						}
 					}
 
 					sQueriesBySId := map[string][]indexerQuery{}
@@ -136,7 +173,7 @@ func InitTorznabIndexerSyncerWorker(conf *WorkerConfig) *Worker {
 					}
 
 					query.SetLimit(-1)
-					if query.IsSupported(tznc.SearchParamIMDBId) {
+					if !nsid.IsAnime && query.IsSupported(tznc.SearchParamIMDBId) {
 						query.Set(tznc.SearchParamIMDBId, nsid.Id)
 						sid := nsid.ToClean()
 						sQuery := indexerQuery{
@@ -159,8 +196,8 @@ func InitTorznabIndexerSyncerWorker(conf *WorkerConfig) *Worker {
 					} else {
 						query.SetT(tznc.FunctionSearch)
 						supportsYear := query.IsSupported(tznc.SearchParamYear)
-						if supportsYear && it.Year != 0 {
-							query.Set(tznc.SearchParamYear, strconv.Itoa(it.Year))
+						if supportsYear && queryMeta.year != 0 {
+							query.Set(tznc.SearchParamYear, strconv.Itoa(queryMeta.year))
 						}
 						for _, title := range queryMeta.titles {
 							var q strings.Builder
@@ -244,6 +281,29 @@ func InitTorznabIndexerSyncerWorker(conf *WorkerConfig) *Worker {
 					}
 
 					log.Debug("indexer search completed", "indexer", client.GetId(), "sid", item.SId, "count", len(results))
+
+					// TODO: download torrent files in a separate queue
+					seenSourceURL := util.NewSet[string]()
+					torzFetchWg := pond.NewPool(5)
+					for i := range results {
+						item := &results[i]
+						if item.HasMissingData() && item.SourceLink != "" {
+							if seenSourceURL.Has(item.SourceLink) {
+								continue
+							}
+							seenSourceURL.Add(item.SourceLink)
+
+							torzFetchWg.Submit(func() {
+								err := item.EnsureMagnet()
+								if err != nil {
+									log.Warn("failed to ensure magnet link for torrent", "error", err)
+								}
+							})
+						}
+					}
+					if err := torzFetchWg.Stop().Wait(); err != nil {
+						log.Warn("errors occurred while fetching torrent magnets", "error", err)
+					}
 
 					tInfosToUpsert := []torrent_info.TorrentItem{}
 					for i := range results {